+++ packages/checkout/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./CheckoutService.js";
+export { CheckoutService } from "./CheckoutService.js";
 export * from "./types.js";